test: handshake lock-degraded snapshot poll on serial test:locks
Move the cross-process status-snapshot gate off the parallel pool.
Wait for the child's first session-refresh.lock EEXIST via a test-only
--import probe that syncs ESM builtin exports, then assert degraded
stale=true/calls=1/no snapshot and a clean calls=2 persisted snapshot.
